+++ Source/Widgets/Dojo/CesiumViewerWidget.js
@@ -26,7 +26,7 @@ define([
         '../../Core/Iso8601',
         '../../Core/Fullscreen',
         '../../Core/computeSunPosition',
-        '../../Core/EventHandler',
+        '../../Core/ScreenSpaceEventHandler',
         '../../Core/FeatureDetection',
         '../../Core/ScreenSpaceEventType',
         '../../Core/Cartesian2',
@@ -81,7 +81,7 @@ define([
         Iso8601,
         Fullscreen,
         computeSunPosition,
-        EventHandler,
+        ScreenSpaceEventHandler,
         FeatureDetection,
         ScreenSpaceEventType,
         Cartesian2,
@@ -662,7 +662,7 @@ define([
             var camera = scene.getCamera();
             camera.position = camera.position.multiplyByScalar(1.5);
 
-            var handler = new EventHandler(canvas);
+            var handler = new ScreenSpaceEventHandler(canvas);
             handler.setInputAction(lang.hitch(this, '_handleLeftClick'), ScreenSpaceEventType.LEFT_CLICK);
             handler.setInputAction(lang.hitch(this, '_handleRightClick'), ScreenSpaceEventType.RIGHT_CLICK);
             handler.setInputAction(lang.hitch(this, '_handleLeftDoubleClick'), ScreenSpaceEventType.LEFT_DOUBLE_CLICK);